chore(examples): Move all code to internal package
